✨Feat(FolderBody): Add 버튼 클릭시 타이틀 및 색상변경
